fix: Basic error handling with requests.

# Conflicts:
#	app/src/main/kotlin/com/example/trainboard/api/Client.kt
#	app/src/main/kotlin/com/example/trainboard/ui/RootView.kt
#	app/src/main/kotlin/com/example/trainboard/ui/StationDropdown.kt
